packages/napari-manual-segmentation/pyproject.toml
[project]
name = "napari-manual-segmentation"
version = "0.1.0"
description = "Adding ."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
  {name = "Tom Blöcker"},
  {email = "Tom.Bloecker@campus.lmu.de"},
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: napari",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "qtpy",
    "napari_toolkit",
    "napari-beacon-layers"
]

[project.entry-points."napari.manifest"]
napari-manual-segmentation = "napari_manual_segmentation:napari.yaml"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]



